Rename reblog_button to boost_button
