+++ app/javascript/mastodon/components/status/boost_button.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import type { StatusVisibility } from '@/mastodon/api_types/statuses';
 import { statusFactoryState } from '@/testing/factories';
 
-import { LegacyReblogButton, StatusBoostButton } from './reblog_button';
+import { LegacyReblogButton, StatusBoostButton } from './boost_button';
 
 interface StoryProps {
   visibility: StatusVisibility;
@@ -13,7 +13,7 @@ interface StoryProps {
 }
 
 const meta = {
-  title: 'Components/Status/ReblogButton',
+  title: 'Components/Status/BoostButton',
   args: {
     visibility: 'public',
     quoteAllowed: true,
