+++ app/javascript/mastodon/components/status/boost_button_utils.ts
